% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_coef.R
\name{vol_coef}
\alias{vol_coef}
\title{Estimates the coefficients for the volatility smile}
\usage{
vol_coef(strike, under_price, du_mat, implicit_vol)
}
\arguments{
\item{strike}{Strike Price.}

\item{under_price}{Underlying asset price.}

\item{du_mat}{Time to maturity.}

\item{implicit_vol}{Observed volatility from trading derivatives of the underlying asset.}
}
\description{
Estimates the coefficients for the volatility smile
}
