% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_point.R
\name{vol_point}
\alias{vol_point}
\title{Estimates the volatility based on the smile curve estimated parameters}
\usage{
vol_point(strike, under_price, du_mat, vol_coef)
}
\arguments{
\item{strike}{Strike Price.}

\item{under_price}{Underlying asset price.}

\item{du_mat}{Time to maturity.}

\item{vol_coef}{Estimated coefficients for the volatility smile curve.}
}
\description{
Estimates the volatility based on the smile curve estimated parameters
}
